added time anomalies

+++ BialHackApi.Base/Interfaces/IMapsService.cs
@@ -5,7 +5,7 @@ namespace BialHackApi.Base.Interfaces
 {
     public interface IMapsService
     {
-        Task<IDistanceDuration> CalculateDistanceByCoords(decimal startLat, decimal startLng, decimal targetLat, decimal targetLng);
+        Task<IDistanceDuration> CalculateDistanceByCoords(double startLat, double startLng, double targetLat, double targetLng);
         Task<ILocation> GetLocationByLatLng(decimal latitude, decimal longitude);
         Task<ILocation> GetLocationByAddress(string address);
         Task<ILocation> GetLocationByExternalId(string externalId);
